More updates
Frame work and everything is so close to being complete, fixed the repeat popup issue after clicking ignore and summarize and score. the manual summarization button is connected and working. for some reason the popup is considered external so its not connecting properyl I need to enable cors? or something like that after that is done the ml part will come into play and i will write the code for it so the dummy numbers will look good. But before that I actually have to write the code to extract the TOS. After I have done both of those things I need to continue to heighten the requirements for TOS shown and then work on UI/UX again. Remember this is personal project solo not team based with 6 plus months